Separate ModuleData and ModuleSchema, introducing Module to combine them
This is definitely good in terms of separating the saving/loading part. However, the fields tend to need access to the schema they're part of... it would be really useful if they "knew" their schema, but that's tricky due to immutability. (The schema could be passed into the conversion methods, on the proviso that it wasn't used until later...) It does make sense for a field to have to be part of a schema, even though it'll make testing more annoying. Hmm.