New command: check instrument defaults
This sets the kick pad on the module to each instrument in turn, setting *only* the instrument field, and then loading all the fields (VEdit etc). It then simulates setting the instrument in the ModuleData, and compares the results. This allows us to check the schema information in terms of instrument defaults.