Add command for turning pages
This currently uses SendKeys, meaning it could only run in .NET Core with the WindowsDesktop SDK.
What I'd like is to be able to run anywhere, but detect "we're running on Windows" and only include the command then. Maybe another day.

For the moment, change to using multi-targeting, and only include the turn-pages command on net472.